IpmiFeaturePkg/GenericIpmi: Move parameter validation
Moves parameter validation in `IpmiSendCommandInternal()` to the
beginning of the function so (1) it is not redundantly checked in
a retry loop and (2) to catch an invalid condition before executing
other function logic.

Also update `IpmiPrintCommand()` to explicitly check for `CommandData`
being `NULL` since it will currently dereference a null pointer if
that pointer variable is `NULL`.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>