Remove the ACPI device scope link.
Remove the ACPI device scope link.

Root cause:

If the system LPC device is not "PC00.LPC0", this SSDT is invaild.

It does not need to be put under any device scope, it just need an ASL device in the root path with all the required methods
then the IPMI device will be initialized success and show up in device manager.