Supports KCS command register declare (#286)
## Description

KCS data and command registers may be decoded as "discontinuous," so a
new PCD is added to manage the command register.

For example:

IO 62h/66h as KCS data and command registers.

For details on how to complete these options and their meaning refer to
[CONTRIBUTING.md](https://github.com/microsoft/mu/blob/HEAD/CONTRIBUTING.md).

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This modification does not affect the original operation.
If the system declares CA2h/CA3h, make sure the IPMI KCS works as
expected.

## Integration Instructions

If your system KCS data and command registers may be decoded as
"discontinuous," please refer this:

Example: 

IO 62h/66h as KCS data and command registers.

```ini
  gIpmiFeaturePkgTokenSpaceGuid.PcdIpmiIoBaseAddress|0x62|UINT16|0xF0000009
  gIpmiFeaturePkgTokenSpaceGuid.PcdIpmiIoCmdRegister|0x66|UINT16|0xF000000A
```

The IPMI device:
Original

![IPMC_ORI](https://github.com/user-attachments/assets/b60b7701-744c-4a52-a419-e1124cb957ea)
Modification

![IPMC_MOD](https://github.com/user-attachments/assets/5f099f2a-335a-4cdf-b11b-83bda4cb79f5)

---------

Co-authored-by: yhsu3 <yhsu3@lenovo.com>
Co-authored-by: Aaron <aaronpop@microsoft.com>