IpmiFeaturePkg: PeiGenericIpmi - Implement RegisterForShadow (#309)
## Description

On platforms that use CAR, when PeiGenericIpmi runs, it will run from
CAR. The PPI Descriptor will be allocated from CAR and installed in the
PEI core.

When permanent memory becomes available, CAR will be torn down, and PPIs
will be transitioned to permemenant memory, but the PPI functions will
still point to the functions that were contained in the flash region.

When MemoryDiscoveredPpiNotifyCallback in CpuMpPei is triggered, the SPI
region will be marked as not present, and any entity that attempts to
access the SPI will trigger a page fault.

Modify the PeiGenericIpmi to register for shadowing (relaunch the entry
point after memory is available) and modify the entry point to detect
that memory is available, and to fixup the PPI descriptor to use the
functions contained in the reloaded PEIM.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested
tested in vendor specific platform that encountered a page fault with
original implementation.

## Integration Instructions
No integration necessary.